Extract ScoreBoard
