+++ game.py
@@ -1,9 +1,7 @@
 """
 Delirious Dentist
 """
-import datetime
 import math
-from operator import itemgetter
 import os
 import random
 import sys
@@ -15,6 +13,7 @@ mymodule_dir = os.path.join(script_dir, "src")
 sys.path.append(mymodule_dir)
 
 import config
+from score_board import ScoreBoard
 from animation.grow import GrowAnimation
 from sprite.dentist import DentistSprite
 from sprite.enemy import EnemySprite
@@ -797,31 +796,6 @@ class InstructionView(arcade.View):
         if key == arcade.key.SPACE:
             self.window.show_view(self.game_view)
 
-
-class ScoreBoard:
-    def store_score(self, score):
-        timestamp = datetime.datetime.now().strftime("%b %d %Y %H:%M:%S")
-        with open(config.HIGH_SCORE_FILE, "a+") as high_score_file:
-            high_score_file.write(f"{score},{timestamp}\n")
-        return timestamp
-
-    def get_high_scores(self, limit=10):
-        with open(config.HIGH_SCORE_FILE, "r+") as high_score_file:
-            lines = list(
-                reversed(
-                    sorted(
-                        map(
-                            lambda x: [int(x[0]), x[1]],
-                            map(
-                                lambda x: x.rstrip().split(","),
-                                high_score_file.readlines(),
-                            ),
-                        ),
-                        key=itemgetter(0),
-                    )
-                )
-            )
-            return lines[0:limit]
 
 def main():
     """Main function"""
